requirements.txt
absl-py==0.13.0
asgiref==3.3.4
attrs==21.2.0
cycler==0.10.0
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
kiwisolver==1.3.1
matplotlib==3.4.2
mediapipe==0.8.6
numpy==1.21.0
opencv-contrib-python==4.5.2.54
Pillow==8.2.0
protobuf==3.17.3
psycopg2==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.2.0
